Added: Permissions for admins Added: Message manager for pop-ups Improved: small fixes & cleanups
TODO: show only incoming events (not all the events in the db)
TODO: Create admin form

+++ panel/index.php
@@ -1,11 +1,13 @@
 <?php
 require_once "assets/php/pager.php";
 require_once "assets/php/database/request.php";
+require_once "../assets/php/messages.php";
 $pager = new PanelPager("Menu");
 
 $pager->setHeader();
 
 $request = new PanelRequest();
+$msg = new Messages();
 
 if ($_SESSION['admin'] == null) {
     header("Location: login.php?error=3");
@@ -15,47 +17,28 @@ $err = filter_input(INPUT_GET, 'error', FILTER_VALIDATE_INT);
 $actu_id = filter_input(INPUT_GET, 'actu-id', FILTER_VALIDATE_INT);
 
 if (isset($_GET['actu_modified'])) {
-    echo "            <div class='pop-up-message' id='pop-up-success'>\n";
-    echo "                <span>Actu modifiée avec succès.</span>\n";
-    echo "            </div>\n";
-}
-
-if (isset($_GET['actu_deleted'])) {
-    echo "            <div class='pop-up-message' id='pop-up-success'>\n";
-    echo "                <span>Actu supprimée avec succès.</span>\n";
-    echo "            </div>\n";
+    $msg->printSuccess("actu_modified");
+} else if (isset($_GET['actu_deleted'])) {
+    $msg->printSuccess("actu_deleted");
 }
 
 if (isset($err)) {
-    switch ($err) {
-        case 2:
-            echo "            <div class='pop-up-message' id='pop-up-fail'>\n";
-            echo "                <span>Les champs ne peuvent pas être vides !</span>\n";
-            echo "            </div>\n";
-            break;
-        case 3:
-            echo "            <div class='pop-up-message' id='pop-up-fail'>\n";
-            echo "                <span>Vous n'êtes pas connecté.</span>\n";
-            echo "            </div>\n";
-            break;
-        case 4:
-            echo "            <div class='pop-up-message' id='pop-up-fail'>\n";
-            echo "                <span>Impossible d'effectuer cette action.</span>\n";
-            echo "            </div>\n";
-    }
+    $msg->printError($err);
 }
 
 if (isset($_GET['logged'])) {
     $username = $_SESSION['admin'];
-
-    echo "            <div class='pop-up-message' id='pop-up-success'>\n";
-    echo "                <span>Connecté avec succès. Bienvenue, $username</span>\n";
-    echo "            </div>\n";
+    $msg->printWelcome($username);
 }
 
 if (isset($_GET['delete-actu'])) {
     if ($actu_id != null) {
-        $request->delete_actu($actu_id);
+        if ($request->isAdmin($_SESSION['admin'])) {
+            $request->delete_actu($actu_id);
+            header("Location: index.php?actu_deleted");
+        } else {
+            header("Location: index.php?error=5");
+        }
     } else {
         header("Location: index.php?error=4");
     }
